refactor: rename packages and restructure project layout
