+++ app/src/androidTest/java/io/github/twyora/douyinenhancer/ExampleInstrumentedTest.kt
@@ -1,12 +1,10 @@
-package com.yst.mkga.hook.dy
+package io.github.twyora.douyinenhancer
 
-import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
-
+import androidx.test.platform.app.InstrumentationRegistry
+import org.junit.Assert
 import org.junit.Test
 import org.junit.runner.RunWith
-
-import org.junit.Assert.*
 
 /**
  * Instrumented test, which will execute on an Android device.
@@ -19,6 +17,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.yst.mkga.hook.dy", appContext.packageName)
+        Assert.assertEquals("io.github.twyora.douyinenhancer.hook", appContext.packageName)
     }
 }
