+++ app/src/main/proto/io/github/twyora/douyinenhancer/hook/configs.proto
@@ -1,8 +1,8 @@
 syntax = "proto3";
 
-package com.yst.mkga.hook.dy.hook;
+package io.github.twyora.douyinenhancer.hook;
 
-option java_package = "com.yst.mkga.hook.dy.hook";
+option java_package = "io.github.twyora.douyinenhancer.hook";
 option java_outer_classname = "Configs";
 option optimize_for = LITE_RUNTIME;
