+++ gradle.properties
@@ -13,10 +13,10 @@ org.gradle.jvmargs=-Xmx2048m -Dfile.encoding=UTF-8
 # org.gradle.parallel=true
 # Kotlin code style for this project: "official" or "obsolete":
 kotlin.code.style=official
-#
-project.namespace=com.yst.mkga.hook.dy
+# Project configuration
+project.namespace=io.github.twyora.douyinenhancer
+project.applicationId=io.github.twyora.douyinenhancer
 project.compileSdk=37
-project.applicationId=com.yst.mkga.hook.dy
 project.minSdk=27
 project.targetSdk=36
 project.versionCode=2
